backend-lm/pyproject.toml
[project]
name = "backend-lm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-multipart==0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
